requirements.txt
Flask
python-dotenv
openai
yt_dlp
assemblyai
